fastapi-medical-app/requirements.txt
fopenai
pandas
numpy
gunicorn
uvicorn
fastapi
pytest
python-multipart
# AI libs
langchain
langchain-openai
langchain-community
scikit-learn
rapidfuzz
aiofiles
# Database
psycopg2-binary
python-dotenv
openpyxl==3.1.2
openai==1.55.0
httpx==0.27.2
python-dotenv==1.0.0
playwright==1.40.0
psutil==5.9.8
rank-bm25==0.2.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
